# src/swagger.yaml
openapi: 3.0.1
info:
  title: Sendit API
  description: >-
    SendIT is a courier service that helps users deliver parcels to different
    destinations.
  version: 1.0.0
servers:
  - url: localhost:3000/v1
...
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: Login Existing user into the system
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: abc@xyz.com
                password:
                  type: string
      responses:
        '200':
          description: logged in
          headers:
            x-auth-token:
              description: authentication token generated by jwt
              schema:
                type: string
                example: >-
                  eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJlbWFpbCI6Im9sdXNlZ3VuQGdtYWlsLmNvbSIsImlzQWRtaW4iOnRydWUsImlhdCI6MTYyOTUwMDgzN30.o1GkKsY8FJdfn3xJlNPDnRudu7Uyxsdpd0YqC-lOGuA
          content:
            application/json:
              schema:
                type: string
                example: logged in
        '400':
          description: Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                  error_code:
                    type: string
                  error_msg:
                    type: string
              examples:
                '400':
                  summary: error - 400
                  value:
                    result: error
                    error_code: 400
                    error_msg: Invalid email or password
                '500':
                  summary: error - 500
                  value:
                    result: error
                    error_code: 500
                    error_msg: Something failed, internal server error
  /auth/signup:
    post:
      tags:
        - auth
      summary: Registers and automatically logs in new user
      operationId: signup User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                lastname:
                  type: string
                email:
                  type: string
                  example: abc@xyz.com
                password:
                  type: string
      responses:
        '200':
          description: signed up
          headers:
            x-auth-token:
              description: authentication token generated by jwt
              schema:
                type: string
                example: >-
                  eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJlbWFpbCI6Im9sdXNlZ3VuQGdtYWlsLmNvbSIsImlzQWRtaW4iOnRydWUsImlhdCI6MTYyOTUwMDgzN30.o1GkKsY8FJdfn3xJlNPDnRudu7Uyxsdpd0YqC-lOGuA
          content:
            application/json:
              schema:
                type: string
                example: signup successful
        '400':
          description: Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                  error_code:
                    type: string
                  error_msg:
                    type: string
              examples:
                '400':
                  summary: error - 400
                  value:
                    result: error
                    error_code: 400
                    error_msg: User already exists or check registration details
                '500':
                  summary: error - 500
                  value:
                    result: error
                    error_code: 500
                    error_msg: Something failed, internal server error
...